working on unit tests
