+++ pyrightconfig.json
@@ -1,12 +1,5 @@
 {
   "include": ["client/valor", "api"],
-  "exclude": [
-    "client/unit-tests",
-    "utils",
-    "api/tests",
-    "examples",
-    "integration_tests",
-    "docs",
-    "migrations"
-  ]
+  "exclude": ["examples", "docs", "migrations"],
+  "python.analysis.extraPaths": ["./client/*", "./api/*"]
 }
